% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aov_batch.R
\name{aov_batch}
\alias{aov_batch}
\title{Title Batch ANOVA Test.}
\usage{
aov_batch(data, y_variable, x_variables, output_file)
}
\arguments{
\item{data}{data which you need to call.}

\item{y_variable}{response factor.}

\item{x_variables}{predictor factors.}

\item{output_file}{The test result will be output as an excel file. This parameter requires you to
write a file name with the suffix "xlsx", such as: "outcome.xlsx"}
}
\description{
The function provides a convenient way to perform batch analysis
of variance (ANOVA) tests on multiple predictor variables with a response
variable. It allows users to conduct hypothesis testing for multiple factors
simultaneously, making it efficient for analyzing experimental or observational
data where the effects of multiple factors need to be examined.
}
\examples{
 my_data <- data.frame(
Group = rep(c("Control", "Treatment"), each = 10),
Age = rep(c("Young", "Middle-aged", "Elderly"), each = 10),
Value = c(22, 18, 20, 24, 21, 23, 25, 19, 20, 22,
         30, 28, 29, 33, 31, 32, 27, 28, 26, 29,
         40, 42, 41, 39, 38, 37, 45, 44, 43, 42))

output <- aov_batch(data = my_data, x_variables = c("Group", "Age"),
          y_variable = "Value","output.xlsx")
}
